+++ broker-core/src/main/java/io/zeebe/broker/system/SystemServiceNames.java
@@ -19,11 +19,15 @@ package io.zeebe.broker.system;
 
 import io.zeebe.broker.system.management.LeaderManagementRequestHandler;
 import io.zeebe.broker.system.metrics.MetricsFileWriter;
+import io.zeebe.broker.system.metrics.MetricsHttpServer;
 import io.zeebe.servicecontainer.ServiceName;
 
 public class SystemServiceNames {
   public static final ServiceName<MetricsFileWriter> METRICS_FILE_WRITER =
       ServiceName.newServiceName("broker.metricsFileWriter", MetricsFileWriter.class);
+
+  public static final ServiceName<MetricsHttpServer> METRICS_HTTP_SERVER =
+      ServiceName.newServiceName("broker.metricsHttpServer", MetricsHttpServer.class);
 
   public static final ServiceName<LeaderManagementRequestHandler>
       LEADER_MANAGEMENT_REQUEST_HANDLER =