+++ broker-core/src/test/java/io/zeebe/broker/test/EmbeddedBrokerConfigurator.java
@@ -88,4 +88,8 @@ public class EmbeddedBrokerConfigurator {
   public static Consumer<BrokerCfg> setSubscriptionApiPort(final int port) {
     return cfg -> cfg.getNetwork().getSubscription().setPort(port);
   }
+
+  public static Consumer<BrokerCfg> setMetricsPort(final int port) {
+    return cfg -> cfg.getMetrics().setPort(port);
+  }
 }